Make SCModalPickerView subclass UIView.
Instead of subclassing UIWindow and ensuring the UIWindow is retained by
the threadDictionary, subclass UIView and create a retain loop between
the SCModalPickerView and the UIWindow. This works and is clean assuming
that the retain loop is broken when the SCModalPickerView is dismissed.

This has the added benefit of keeping threadDictionary clean and not
breaking if the user tries to display two SCModalPickerViews at the same
time.